+++ src/app/(home)/(posts)/_components/forum-detail/InputComment.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useQuery } from '@tanstack/react-query';
+
 import MDEditor, { commands } from '@uiw/react-md-editor';
 import { useState } from 'react';
