+++ src/components/timeForToday.ts
@@ -1,7 +1,3 @@
-import { toast } from 'react-toastify';
-
-//공용으로 사용하는 함수 넣어두는게 좋을 것같아서 만들었습니다. 가져다가 쓰시면 됩니다.
-
 // 현재 시간 기준으로 시간 표시 ex) 방금전, 30분전, 1시간전 등..
 // {timeForToday(post.created_at)} <- 이런식으로 사용 하시면 됩니다.
 export const timeForToday = (value: string) => {
@@ -25,17 +21,4 @@ export const timeForToday = (value: string) => {
   }
 
   return `${Math.floor(betweenTimeDay / 365)}년전`;
-};
-
-//링크 공유
-//handleRinkCoy(들어갈 주소 입력)
-//ex) onClick={() => handleRinkCopy(`${process.env.NEXT_PUBLIC_BASE_URL}/forum/${forum.id}`)}
-
-export const handleRinkCopy = async (text: string) => {
-  try {
-    await navigator.clipboard.writeText(text);
-    toast.success('링크가 복사되었습니다.', {
-      autoClose: 2000
-    });
-  } catch (error) {}
 };
